fix: refatorei o list
