+++ src/components/list/style.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const List = styled.ul`
+  margin: 16px auto;
+  max-width: 1100px;
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: center;
+  gap: 16px;
+`;
